==> 01-ETL.ipynb <==
import os
import pandas as pd
import numpy as np
from IPython.display import display

cwd = os.getcwd()
folder = os.path.join(cwd, "Citibike_Data")

# Citibike New York City Data

- One can find the system-data at the following [link](https://www.citibikenyc.com/system-data)
    - [S3 Amazon Bucket](https://s3.amazonaws.com/tripdata/index.html)

# get all filenames for the .csv.zip files
for (dirpath, dirnames, filenames) in os.walk(folder):
    pass
filenames = sorted(filenames)
filenames[:5]

# Workhorse Function to read in the 36 datafiles

- files are compressed and presented as .csv.zip

# function that is the workhouse to generate the entire dataset
def Get_Data(filename=filenames[0], pd=pd):
    """
    Does all the work.  Read in DataFrames, parse them then return the dataframes
    """
    # read into dataframe
    df = pd.read_csv(os.path.join(folder, filename), compression="zip").iloc[:, 1:]

    # change to datetime both start time and stop time
    # columns change their names for some reason...
    df.iloc[:, 0] = pd.to_datetime(df.iloc[:, 0])
    df.iloc[:, 1] = pd.to_datetime(df.iloc[:, 1])

    # get the Year, Month and Day
    df["Start Year"] = df.iloc[:, 0].dt.year
    df["Start Month"] = df.iloc[:, 0].dt.month
    df["Start Day"] = df.iloc[:, 0].dt.day
    df["Start Hour"] = df.iloc[:, 0].dt.hour
    df["Duration_Seconds"] = (df.iloc[:, 1] - df.iloc[:, 0]).dt.total_seconds().astype('int64')

    # no longer need Start Time and Stop Time
    df = df.iloc[:,2:]
    
    # make sure column names are correct
    df.columns = ['Start Station ID', 'Start Station Name', 'Start Station Latitude',
                  'Start Station Longitude', 'End Station ID', 'End Station Name',
                  'End Station Latitude', 'End Station Longitude', 'Bike ID', 'User Type',
                  'Birth Year', 'Gender', 'Start Year', 'Start Month', 'Start Day', 'Start Hour',
                  'Duration_Seconds']
    
    # return the dataframe
    return df

# for each file name append the dataframes together
for x in filenames:
    
    # for first one just generate the dataframe
    if x == filenames[0]:
        df_total = Get_Data(filename=x)
    
    # for the following append them
    df_total = df_total.append(Get_Data(filename=x))

# reset the index
df_total.reset_index(drop=True, inplace=True)
    
# turn everything categorical    
df_total = df_total.astype({"Start Station ID":'category',
                 "Start Station Name":'category',
                 "End Station ID":'category',
                 "End Station Name":'category',
                 "Bike ID":'category',
                 "User Type":'category',
                 "Birth Year":'category',
                 "Gender":'category',
                 "Start Year":'category',
                 "Start Month":'category',
                 "Start Day":'category',
                 'Start Hour':'category'})
df_total.info()

df_total.sample(10)

# Generate a Lookup table for all the Station ID's

- store the Start and End station ID's in the table but remove the rest of the information
- save the geospatial information in a .csv file

# generate a table of all unique Start and End Stations
df_unique_start = df_total.loc[:, ['Start Station ID', 'Start Station Name',
                                   'Start Station Latitude', 'Start Station Longitude']].copy()
df_unique_end = df_total.loc[:, ['End Station ID', 'End Station Name',
                                   'End Station Latitude', 'End Station Longitude']].copy()

# now slice out all the Station Stuff except for the ID's
df_total = df_total.loc[:, ['Start Station ID', 'End Station ID', 'Bike ID', 'User Type',
       'Birth Year', 'Gender', 'Start Year', 'Start Month', 'Start Day', 'Start Hour',
       'Duration_Seconds']]

df_unique_start.head()

# drop all duplicates
df_unique_start = df_unique_start.drop_duplicates()
df_unique_end = df_unique_end.drop_duplicates()
df_unique_end.columns = df_unique_start.columns

df_unique_start.shape, df_unique_end.shape

# append the 2 lists and drop al duplicates as well by ID
df_unique = df_unique_start.append(df_unique_end)
df_unique = df_unique.drop_duplicates(subset=["Start Station ID"]).dropna(subset=["Start Station ID"])
df_unique.shape

# write the unique values to an external lookup table
df_unique.sort_values(by=["Start Station ID"], inplace=True)
df_unique.to_csv(os.path.join(cwd, "Citibike_Clean", "Station_ID.csv"),
                 index=False)

df_total.info()

# The main dataset for the 3 year period is written into memory

- January 2017 - December 2019 is the date range

# write the csv
df_total.to_csv(os.path.join(cwd, "Citibike_Clean", "CitiBike-NYC.csv.zip"),
                index=False, compression="zip")

df_total.head()

# to load the dataset back into memory
del df_total

# just to show you how
df_total = pd.read_csv(os.path.join(cwd, "Citibike_Clean", "CitiBike-NYC.csv.zip"),
                       compression="zip")
df_total.head()

df_total.info()

del df_total
