==> 02-ETL_2.ipynb <==
import os
import pandas as pd
import numpy as np
from IPython.display import display

cwd = os.getcwd()

# load dataset into memory
# drop the stations that do not have an ID as distance cannot be determined
df = pd.read_csv(os.path.join(cwd, "Citibike_Clean", "CitiBike-NYC.csv.zip"),
                       compression="zip").dropna(subset=["Start Station ID", "End Station ID"])

# change the datatype of Start and End Station ID to integer
# all nulls have been dropped
df["Start Station ID"] = df["Start Station ID"].astype("int64")
df["End Station ID"] = df["End Station ID"].astype("int64")

df.head()

# Categorical Data Type used for RAM memory savings

- pandas Categorical data type utilized

# check out the memory size of the pandas dataframe
print(f"{df.memory_usage().sum()/1_000_000} Megabytes")
df.info()

# change the datatype for several columns into categorical to save on RAM memory
df["Start Station ID"] = pd.Categorical(df["Start Station ID"])
df["End Station ID"] = pd.Categorical(df["End Station ID"])
df["Bike ID"] = pd.Categorical(df["Bike ID"])
df["User Type"] = pd.Categorical(df["User Type"])
df["Birth Year"] = pd.Categorical(df["Birth Year"])
df["Gender"] = pd.Categorical(df["Gender"])
df["Start Year"] = pd.Categorical(df["Start Year"])
df["Start Month"] = pd.Categorical(df["Start Month"])
df["Start Day"] = pd.Categorical(df["Start Day"])
df["Start Hour"] = pd.Categorical(df["Start Hour"])

# notice the huge memory reductions
print(f"{df.memory_usage().sum()/1_000_000} Megabytes")
df.info()

# Remove any null Station ID's from the dataset

- write the new dataset into memory

# write the intermediate stage back into memory, write the csv
df.to_csv(os.path.join(cwd, "Citibike_Clean", "CitiBike-NYC-no_nulls.csv.zip"),
                index=False, compression="zip")

# free up the variable df
del df

# read the new dataframe back into memory
df = pd.read_csv(os.path.join(cwd, "Citibike_Clean", "CitiBike-NYC-no_nulls.csv.zip"),
                       compression="zip")

# change the datatype for several columns into categorical to save on RAM memory
df["Start Station ID"] = pd.Categorical(df["Start Station ID"])
df["End Station ID"] = pd.Categorical(df["End Station ID"])
df["Bike ID"] = pd.Categorical(df["Bike ID"])
df["User Type"] = pd.Categorical(df["User Type"])
df["Birth Year"] = pd.Categorical(df["Birth Year"])
df["Gender"] = pd.Categorical(df["Gender"])
df["Start Year"] = pd.Categorical(df["Start Year"])
df["Start Month"] = pd.Categorical(df["Start Month"])
df["Start Day"] = pd.Categorical(df["Start Day"])
df["Start Hour"] = pd.Categorical(df["Start Hour"])

df.head()

# Merge the Station ID lookup table to the main dataset

- obtain latitude and longitudes for the Station IDs

# load the location dataset into memory
df_loc = pd.read_csv(os.path.join(cwd, "Citibike_Clean", "Station_ID.csv")).dropna()

# change datatype for Start Station ID to integer then category
df_loc["Start Station ID"] = df_loc.astype({"Start Station ID": "int64"})
df_loc["Start Station ID"] = pd.Categorical(df_loc["Start Station ID"])

df_loc.head()

# need to merge the 2 datasets twice over to get
# Start Station and End Station Latitudes and Longitudes
df_total = df.merge(df_loc, left_on="Start Station ID",
                    right_on="Start Station ID", suffixes=("", "_"))

# free up memory by deleting objects and columns
del df
df_total.drop(columns=["Start Station Name"], inplace=True)

# merge again to get further latitudes and longitudes for the End Stations
df_total = df_total.merge(df_loc, left_on="End Station ID",
                    right_on="Start Station ID", suffixes=("", "_"))

# free up memory by deleting objects and columns
del df_loc
df_total.drop(columns=["Start Station Name", "Start Station ID_"], inplace=True)

df_total.rename(columns={"Start Station Latitude_": "End Station Latitude",
                          "Start Station Longitude_": "End Station Longitude"}, inplace=True)
df_total.head()

# Calculate absolute distance between pairs of latitudes and longitudes

- use the [haversine formula](https://en.wikipedia.org/wiki/Haversine_formula)
    - Numpy optimized version was created below

# Python Haversine formula for numpy arrays
def np_haversine(latit_1=-90, latit_2=90, longit_1=0, longit_2=180, radius=6371000):
    """
    Calculate great circle distance between 2 points on a sphere (Earth is default).
    Positions are in decimal degrees.
    Function was rewritten to be calculated unsing Numpy methods.
    Returns value in metres (m).
    """
    # convert decimal degrees to radians
    latit_1 = np.radians(latit_1)
    latit_2 = np.radians(latit_2)
    longit_1 = np.radians(longit_1)
    longit_2 = np.radians(longit_2)
    
    # calculate the difference between latitudes and longitudes
    delta_latitude = latit_1 - latit_2
    delta_longitude = longit_1 - longit_2
    
    # great circle distance
    a = np.sin(delta_latitude / 2)**2 + np.cos(latit_1) * np.cos(latit_2) * np.sin(delta_longitude / 2)**2
    c = 2 * np.arcsin(np.sqrt(a))
    
    # return the distance
    return radius * c

np_haversine()/6371000

# use the haversine formula to calculate the distance between
# stations (unit of metres) and add 1 metre to deal with infinity values from division
df_total["Station_Distance"] = pd.Series(np.around(np_haversine(latit_1=df_total["Start Station Latitude"].values,
                                            latit_2=df_total["End Station Latitude"].values,
                                            longit_1=df_total["Start Station Longitude"].values,
                                            longit_2=df_total["End Station Longitude"].values)
                                        , decimals=0)).astype("int64") + 1

# remove the Latitudes and Longitudes as they are no longer needed
df_total.drop(columns=["Start Station Latitude",
                       "Start Station Longitude",
                       "End Station Latitude",
                       "End Station Longitude"], inplace=True)

# calculate seconds travelled per metre
df_total["Seconds_per_Metre"] = (df_total["Duration_Seconds"]/df_total["Station_Distance"]).round(decimals=2)
df_total.head()

# Write the new dataset with the distance between Start and End Stations

- is start and end station are the same the answer will be 1 (to prevent any divide by 0 errors later on)
    - otherwise the distance is calculate into metres

# write the final dataset back into memory, write the csv
df_total.to_csv(os.path.join(cwd, "Citibike_Clean", "CitiBike-NYC-FINAL.csv.zip"),
                index=False, compression="zip")
del df_total

# read the new dataframe back into memory
df = pd.read_csv(os.path.join(cwd, "Citibike_Clean", "CitiBike-NYC-FINAL.csv.zip"),
                       compression="zip")

# change the datatype for several columns into categorical to save on RAM memory
df["Start Station ID"] = pd.Categorical(df["Start Station ID"])
df["End Station ID"] = pd.Categorical(df["End Station ID"])
df["Bike ID"] = pd.Categorical(df["Bike ID"])
df["User Type"] = pd.Categorical(df["User Type"])
df["Birth Year"] = pd.Categorical(df["Birth Year"])
df["Gender"] = pd.Categorical(df["Gender"])
df["Start Year"] = pd.Categorical(df["Start Year"])
df["Start Month"] = pd.Categorical(df["Start Month"])
df["Start Day"] = pd.Categorical(df["Start Day"])
df["Start Hour"] = pd.Categorical(df["Start Hour"])

df.head()

# check to see if there are any infinite values in Seconds_per_Metre
# looks empty
df[~np.isfinite(df["Seconds_per_Metre"])]

# clean up memory
del df
