==> 03-Data_Aggregation.ipynb <==
import os
import pandas as pd
import numpy as np
from IPython.display import display

cwd = os.getcwd()

# read the new dataframe back into memory
df = pd.read_csv(os.path.join(cwd, "Citibike_Clean", "CitiBike-NYC-FINAL.csv.zip"),
                       compression="zip")

# change the datatype for several columns into categorical to save on RAM memory
df["Start Station ID"] = pd.Categorical(df["Start Station ID"])
df["End Station ID"] = pd.Categorical(df["End Station ID"])
df["Bike ID"] = pd.Categorical(df["Bike ID"])
df["User Type"] = pd.Categorical(df["User Type"])
df["Birth Year"] = pd.Categorical(df["Birth Year"])
df["Gender"] = pd.Categorical(df["Gender"])
df["Start Year"] = pd.Categorical(df["Start Year"])
df["Start Month"] = pd.Categorical(df["Start Month"])
df["Start Day"] = pd.Categorical(df["Start Day"])
df["Start Hour"] = pd.Categorical(df["Start Hour"])

df.head()

# First Aggregation

- aggregate on Year, Month, Day and Hour
    - sum and count on Duration and Distance

# aggregate on year, month, day and hour for duration in seconds
df_agg = (df.loc[:, ['Start Year', 'Start Month', 'Start Day',
           'Start Hour', 'Duration_Seconds', 'Station_Distance']]
   .groupby(['Start Year', 'Start Month', 
             'Start Day', 'Start Hour']).agg(["sum", "count"]))
df_agg.columns = df_agg.columns.droplevel()
df_agg.columns = ["Duration_Seconds_SUM", "Duration_Seconds_COUNT",
                  "Station_Distance_SUM", "Station_Distance_COUNT"]
df_agg.reset_index(inplace=True)

# fill empty values with 0
df_agg.fillna(value={
    "Duration_Seconds_SUM":0,
    "Duration_Seconds_COUNT":0,
    "Station_Distance_SUM":1,
    "Station_Distance_COUNT":1
                    }, inplace=True)

# set the datatype for sum and count
df_agg["Duration_Seconds_SUM"] = df_agg["Duration_Seconds_SUM"].astype("int64")
df_agg["Duration_Seconds_COUNT"] = df_agg["Duration_Seconds_COUNT"].astype("int64")
df_agg["Station_Distance_SUM"] = df_agg["Station_Distance_SUM"].astype("int64")
df_agg["Station_Distance_COUNT"] = df_agg["Station_Distance_COUNT"].astype("int64")

# write this to csv file
df_agg.to_csv(os.path.join(cwd, "Citibike_Aggregate", "GroupBy_Year_Month_Day_Hour_Seconds_Distance.csv"),
                 index=False)

df_agg.head()

# Second Aggregation

- aggregate on year, month, day, hour , user type and gender
    - sum and count on duration and distance

del df_agg

# year, month, day, and hour, user type, gender groupby
df_agg = (df.loc[:, ['Start Year', 'Start Month', 'Start Day',
           'Start Hour', 'User Type', 'Gender', 'Duration_Seconds', 'Station_Distance']]
   .groupby(['Start Year', 'Start Month', 
             'Start Day', 'Start Hour',
             'User Type', 'Gender']).agg(["sum", "count"]))
df_agg.columns = df_agg.columns.droplevel()
df_agg.columns = ["Duration_Seconds_SUM", "Duration_Seconds_COUNT",
                  "Station_Distance_SUM", "Station_Distance_COUNT"]
df_agg.reset_index(inplace=True)

# fill empty values with 0
df_agg.fillna(value={
    "Duration_Seconds_SUM":0,
    "Duration_Seconds_COUNT":0,
    "Station_Distance_SUM":1,
    "Station_Distance_COUNT":1
                    }, inplace=True)

# set the datatype for sum and count
df_agg["Duration_Seconds_SUM"] = df_agg["Duration_Seconds_SUM"].astype("int64")
df_agg["Duration_Seconds_COUNT"] = df_agg["Duration_Seconds_COUNT"].astype("int64")
df_agg["Station_Distance_SUM"] = df_agg["Station_Distance_SUM"].astype("int64")
df_agg["Station_Distance_COUNT"] = df_agg["Station_Distance_COUNT"].astype("int64")

# write this to csv file
df_agg.to_csv(os.path.join(cwd, "Citibike_Aggregate", "GroupBy_Year_Month_Day_Hour_User_Gender_Seconds_Distance.csv"),
                 index=False)

df_agg.head()

# Third Aggregation

- aggregate on Bike ID only
    - sum and count on duration and station

del df_agg

# bike id groupby
df_agg = (df.loc[:, ["Bike ID", 'Duration_Seconds', 'Station_Distance']]
   .groupby(['Bike ID']).agg(["sum", "count"]))
df_agg.columns = df_agg.columns.droplevel()
df_agg.columns = ["Duration_Seconds_SUM", "Duration_Seconds_COUNT",
                  "Station_Distance_SUM", "Station_Distance_COUNT"]
df_agg.reset_index(inplace=True)

# fill empty values with 0
df_agg.fillna(value={
    "Duration_Seconds_SUM":0,
    "Duration_Seconds_COUNT":0,
    "Station_Distance_SUM":1,
    "Station_Distance_COUNT":1
                    }, inplace=True)

# set the datatype for sum and count
df_agg["Duration_Seconds_SUM"] = df_agg["Duration_Seconds_SUM"].astype("int64")
df_agg["Duration_Seconds_COUNT"] = df_agg["Duration_Seconds_COUNT"].astype("int64")
df_agg["Station_Distance_SUM"] = df_agg["Station_Distance_SUM"].astype("int64")
df_agg["Station_Distance_COUNT"] = df_agg["Station_Distance_COUNT"].astype("int64")

# sort the aggregation so the largest values are on top
df_agg.sort_values(by=['Duration_Seconds_SUM'], ascending=False, inplace=True)
df_agg.reset_index(drop=True, inplace=True)

# write this to csv file
df_agg.to_csv(os.path.join(cwd, "Citibike_Aggregate", "GroupBy_bikeid_Seconds_Distance.csv"),
                 index=False)

# grab the first 50 bike ID then slice the dataframe down
df_top_50_bikes = df_agg["Bike ID"][:50]
df_top_50 = df.loc[(df["Bike ID"].cat.codes.isin(df_top_50_bikes.cat.codes)), :].copy()
df_top_50["Bike ID"] = df_top_50["Bike ID"].astype("int64")

df_agg.head()

# make sure the datatype is correct (no residual memory of other bikes)
df_top_50.info()

# verify only top 50 bikes are present
len(df_top_50['Bike ID'].unique())

# Fourth Aggregation

- groupby on year, month, day, hour, bike ID for only the top 50 bikes
    - sum and count on duration and distance

del df_agg, df

# year, month, day, and hour, top 50 bike id groupby
df_agg = (df_top_50.loc[:, ["Start Year", "Start Month", "Start Day", "Start Hour",
                            "Bike ID", 'Duration_Seconds', 'Station_Distance']]
   .groupby(["Start Year", "Start Month", "Start Day", "Start Hour",
             'Bike ID']).agg(["sum", "count"]))
df_agg.columns = df_agg.columns.droplevel()
df_agg.columns = ["Duration_Seconds_SUM", "Duration_Seconds_COUNT",
                  "Station_Distance_SUM", "Station_Distance_COUNT"]
df_agg.reset_index(inplace=True)

# fill empty values with 0
df_agg.fillna(value={
    "Duration_Seconds_SUM":0,
    "Duration_Seconds_COUNT":0,
    "Station_Distance_SUM":1,
    "Station_Distance_COUNT":1
                    }, inplace=True)

# set the datatype for sum and count
df_agg["Duration_Seconds_SUM"] = df_agg["Duration_Seconds_SUM"].astype("int64")
df_agg["Duration_Seconds_COUNT"] = df_agg["Duration_Seconds_COUNT"].astype("int64")
df_agg["Station_Distance_SUM"] = df_agg["Station_Distance_SUM"].astype("int64")
df_agg["Station_Distance_COUNT"] = df_agg["Station_Distance_COUNT"].astype("int64")

# write this to csv file
df_agg.to_csv(os.path.join(cwd, "Citibike_Aggregate", "GroupBy_Year_Month_Day_Hour_top50_bikeid_Seconds_Distance.csv"),
                 index=False)

display(df_agg.head())

del df_agg
