==> 08-Statistical_Test.ipynb <==
import os
import pandas as pd
import numpy as np
from IPython.display import display

cwd = os.getcwd()

# read in preaggregated csv file
df_agg = pd.read_csv(os.path.join(cwd,
                          "Citibike_Aggregate",
                          "GroupBy_Year_Month_Day_Hour_User_Gender_Seconds_Distance.csv"))

df_agg.head()

df_agg_gender = df_agg.drop(['Start Year','Start Month','Start Day','Start Hour', 'User Type','Duration_Seconds_COUNT','Station_Distance_SUM','Station_Distance_COUNT'], axis = 1)

df_agg_gender

## Import the packages
from scipy import stats
import random

#Sampling
#Women's duration seconds list

df_agg_women = df_agg_gender[(df_agg_gender.Gender == 1)]

women_bike_duration_list_all = df_agg_women['Duration_Seconds_SUM'].to_list()

women_bike_duration_list = random.sample(women_bike_duration_list_all, 1000)

#print(women_bike_duration_list)

#Men's duration seconds list

df_agg_men = df_agg_gender[(df_agg_gender.Gender == 2)]

men_bike_duration_list_all = df_agg_men['Duration_Seconds_SUM'].to_list()

men_bike_duration_list = random.sample(men_bike_duration_list_all, 1000)

#print(men_bike_duration_list)

# calculate means and standard deviations

mean_women = np.mean(women_bike_duration_list)
mean_men = np.mean(men_bike_duration_list)

#print(mean_women, mean_men)

print({"women mean":mean_women,"men mean":mean_men})

#calculate standard deviations
std_women = np.std(women_bike_duration_list)
std_men = np.std(men_bike_duration_list)

print({"women standard deviation":std_women, "men standard deviation":std_men})

# calculate standard errors

n1 = len(women_bike_duration_list)
n2 = len(men_bike_duration_list)

se_women = std_women/np.sqrt(n1)
se_men = std_men/np.sqrt(n2)

print({"women standard error":se_women, "men standard error":se_men})

# standard error on the difference between the samples
sed = np.sqrt(se_women**2.0 + se_men**2.0)

sed

# calculate the t statistic
t_stat = (mean_women - mean_men) / sed

print({"t value for duration seconds":t_stat})

#station distance:

df_agg_gender_dis = df_agg.drop(['Start Year','Start Month','Start Day','Start Hour', 'User Type','Duration_Seconds_COUNT','Duration_Seconds_SUM','Station_Distance_COUNT'], axis = 1)

df_agg_gender_dis.head()

## Import the packages
from scipy import stats

#Sampling
#Women's station distance list

df_agg_women_dis = df_agg_gender_dis[(df_agg_gender.Gender == 1)]

women_station_dis_list_all = df_agg_women_dis['Station_Distance_SUM'].to_list()

women_station_dis_list = random.sample(women_station_dis_list_all,1000)

#print(women_station_dis_list)

#Women's station distance list

df_agg_men_dis = df_agg_gender_dis[(df_agg_gender.Gender == 2)]

men_station_dis_list_all = df_agg_men_dis['Station_Distance_SUM'].to_list()

men_station_dis_list = random.sample(men_station_dis_list_all,1000)

#print(men_station_dis_list)

# calculate means and standard deviations

mean_women_dis = np.mean(women_station_dis_list)
mean_men_dis = np.mean(men_station_dis_list)


print({"women mean dis":mean_women_dis,"men mean dis":mean_men_dis})

#calculate standard deviations
std_women_dis = np.std(women_station_dis_list)
std_men_dis = np.std(men_station_dis_list)

print({"women standard deviation dis":std_women_dis, "men standard deviation":std_men_dis})

# calculate standard errors

n3 = len(women_station_dis_list)
n4 = len(men_station_dis_list)

se_women_dis = std_women_dis/np.sqrt(n3)
se_men_dis = std_men_dis/np.sqrt(n4)

print({"women standard error dis":se_women_dis, "men standard error dis":se_men_dis})

# standard error on the difference between the samples
sed_dis = np.sqrt(se_women_dis**2.0 + se_men_dis**2.0)

sed_dis

# calculate the t statistic
t_stat_dis = (mean_women_dis - mean_men_dis) / sed_dis

t_stat_dis

print({"t value for station distance":t_stat_dis})

#Conclusion

#Null Hypothesis: The data for duration time periods and distances between stations have the same distribution regardless of the gender. 
#Alternative Hypothesis: The data for duration time periods and distances between stations have different distributions for women and men.

#The null hypothesis is rejcted for both datasets because both T values are much larger than 0.
